requirements.txt
apache-airflow
datasets
pandas
numpy
matplotlib
scikit-learn
joblib
prefect
fastapi
uvicorn
requests
